requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
Flask
tensorflow
streamlit
keras
pillow
tempfile
#-e .
